feat: shows TagButtonChunkFixed in app
