+++ src/components/dialog_fixed_tags/index.tsx
@@ -11,20 +11,19 @@ import {
   isFixedTagsDialogOpenState,
   fixedTagsState,
 } from '@/recoil/words/words.state'
-import TagChipDeletable from '../atom_tag_chip/index.deletable'
 import StyledTextField from '@/atoms/StyledTextField'
 import { useKeyPress } from '@/hooks/use-key-press.hook'
+import TagButtonChunkFixed from '../molecule_tag_button_chunk/index.fixed'
 
 const FixedTagsDialog: FC = () => {
   const [input, setInput] = useState(``)
   const fixedTagsDialogOpen = useRecoilValue(isFixedTagsDialogOpenState)
-  const fixedTags = useRecoilValue(fixedTagsState)
   const onClose = useResetRecoilState(isFixedTagsDialogOpenState)
 
   const onHitEnter = useRecoilCallback(
-    ({ set }) =>
+    ({ set, snapshot }) =>
       async () => {
-        const fixedTagSet = new Set(fixedTags)
+        const fixedTagSet = new Set(await snapshot.getPromise(fixedTagsState))
         const trimmed = input.trim()
         if (trimmed.length === 0) return
 
@@ -32,20 +31,10 @@ const FixedTagsDialog: FC = () => {
         set(fixedTagsState, Array.from(fixedTagSet))
         setInput(``) // input is cleared after adding a tag
       },
-    [input, fixedTags, setInput],
+    [input, setInput],
   )
 
   useKeyPress(onHitEnter, `Enter`)
-
-  const onClickDelete = useRecoilCallback(
-    ({ set }) =>
-      async (label: string) => {
-        const fixedTagSet = new Set(fixedTags)
-        fixedTagSet.delete(label)
-        set(fixedTagsState, Array.from(fixedTagSet))
-      },
-    [fixedTags],
-  )
 
   if (!fixedTagsDialogOpen) return null
 
@@ -64,9 +53,7 @@ const FixedTagsDialog: FC = () => {
           label={`Insert your Fixed Tags here`}
         />
         <Box pb={1} />
-        {fixedTags.map((tag) => (
-          <TagChipDeletable key={tag} label={tag} onClick={onClickDelete} />
-        ))}
+        <TagButtonChunkFixed />
       </DialogContent>
     </StyledDialog>
   )
